### 2025-03-20 eb8b32f
docs: quote the affiliation as in the template
