### 2025-03-25 5ed09d8
build: add the frontend docker build
