+++ .github/workflows/frontend-docker.yaml
@@ -1,0 +1,104 @@
+name: Frontend Nuxt Docker Build
+
+on:
+  push:
+    branches: [ main, dev ]
+    paths:
+      - 'frontend/**'
+      - '.github/workflows/frontend-docker.yaml'
+  pull_request:
+    branches: [ main, dev ]
+    paths:
+      - 'frontend/**'
+      - '.github/workflows/frontend-docker.yaml'
+  workflow_dispatch:
+
+concurrency:
+  group: ${{ github.workflow }}-${{ github.ref }}
+  cancel-in-progress: true
+
+permissions:
+  contents: read
+  packages: write
+
+jobs:
+  build-nuxt-docker:
+    name: Build Nuxt Frontend Image
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v4
+
+      - name: Free up disk space
+        run: |
+          sudo apt-get remove -y '^ghc-.*' '^dotnet-.*' '^llvm-.*' '^mono-.*' '^php.*' '^mongodb-.*' '^mysql-.*'
+          sudo apt-get autoremove -y
+          sudo apt-get clean
+          sudo rm -rf /usr/share/dotnet /usr/local/lib/android /opt/ghc /opt/hostedtoolcache/CodeQL
+          df -h
+
+      - name: Login to GitHub Container Registry
+        uses: docker/login-action@v3
+        with:
+          registry: ghcr.io
+          username: ${{ github.actor }}
+          password: ${{ secrets.GITHUB_TOKEN }}
+
+      - name: Set up Docker Buildx
+        uses: docker/setup-buildx-action@v3
+
+      - name: Extract metadata for Docker
+        id: meta
+        uses: docker/metadata-action@v5
+        with:
+          images: ghcr.io/${{ github.repository }}/frontend
+          tags: |
+            type=ref,event=branch
+            type=ref,event=pr
+            type=sha,format=short
+            type=raw,value=latest,enable=${{ github.ref == format('refs/heads/{0}', 'main') }}
+
+      - name: Build and push Nuxt Docker image
+        uses: docker/build-push-action@v5
+        with:
+          context: ./frontend
+          file: ./frontend/Dockerfile
+          push: ${{ github.event_name != 'pull_request' || github.event.pull_request.head.repo.full_name == github.repository }}
+          tags: ${{ steps.meta.outputs.tags }}
+          labels: ${{ steps.meta.outputs.labels }}
+          cache-from: type=gha,scope=frontend-pnpm
+          cache-to: type=gha,mode=max,scope=frontend-pnpm
+          build-args: |
+            NODE_ENV=production
+          platforms: linux/amd64
+
+      - name: Test the built Nuxt image
+        if: github.event_name != 'pull_request' || github.event.pull_request.head.repo.full_name == github.repository
+        run: |
+          # Pull the built image (first tag only to avoid issues)
+          FIRST_TAG=$(echo "${{ steps.meta.outputs.tags }}" | head -n 1)
+          docker pull $FIRST_TAG
+          
+          # Run the container in the background
+          CONTAINER_ID=$(docker run -d -p 3000:3000 $FIRST_TAG)
+          
+          # Wait for the app to start
+          sleep 10
+          
+          # Test if the server is running
+          if curl -s -f http://localhost:3000 > /dev/null; then
+            echo "✅ Nuxt application is running successfully"
+          else
+            echo "❌ Failed to access Nuxt application"
+            docker logs $CONTAINER_ID
+            docker stop $CONTAINER_ID
+            exit 1
+          fi
+          
+          # Clean up
+          docker stop $CONTAINER_ID
+          
+      - name: Output Image URI
[…]
