### 2025-03-27 15f68ff
feat: decouple atlas generation from metadata existance
