### 2025-03-27 e1b74a5
feat: fetch all files via a server route
resolves issues if run locally, docker, or singularity on an HPC
