### 2025-03-27 b260796
build: :green_heart: add the version to the frontend docker image
