### 2025-03-27 bbf5045
test: update test cases
