### 2025-03-28 dabb4a1
feat: add size constraint to atlas generation and improve compatibility
