+++ frontend/server/api/atlasGenerator.ts
@@ -1,3 +1,4 @@
+import type { Sharp } from 'sharp'
 import fs from 'node:fs'
 import path from 'node:path'
 import process from 'node:process'
@@ -5,6 +6,15 @@ import { useStorage } from '#imports'
 import { defineEventHandler } from 'h3'
 import sharp from 'sharp'
 
+// Constants
+const MAX_ATLAS_WIDTH = 8192
+const MAX_ATLAS_HEIGHT = 8192
+const MAX_PIXEL_COUNT = 67108864
+const METADATA_BATCH_SIZE = 2
+const RESIZE_BATCH_SIZE = 2
+const ALLOWED_EXTENSIONS = ['.png', '.jpg', '.jpeg']
+
+// Interfaces
 interface AtlasStatus {
   status: string
   progress: number
@@ -12,7 +22,121 @@ interface AtlasStatus {
   lastUpdated: string
 }
 
-async function updateStatus(progress: number, message: string, status: string = 'in_progress'): Promise<void> {
+interface ImageMeta {
+  filename: string
+  width: number
+  height: number
+  originalWidth: number
+  originalHeight: number
+}
+
+interface PackedImage extends ImageMeta {
+  x: number
+  y: number
+}
+
+interface Rect {
+  x: number
+  y: number
+  width: number
+  height: number
+}
+
+// Bin-packing algorithm
+class Packer {
+  private bins: Rect[] = []
+  private readonly maxWidth: number
+  private currentHeight: number = 0
+
+  constructor(maxWidth: number) {
+    this.maxWidth = maxWidth
+    this.bins.push({ x: 0, y: 0, width: maxWidth, height: 0 })
+  }
+
+  private findBestBin(width: number, height: number): Rect | null {
+    let bestBin: Rect | null = null
+    let bestScore = Infinity
+
+    for (const bin of this.bins) {
+      if (bin.width < width || bin.height < height)
+        continue
+
+      const score = bin.width - width
+      if (score < bestScore) {
+        bestScore = score
+        bestBin = bin
+      }
+    }
+
+    return bestBin
+  }
+
+  pack(image: ImageMeta): PackedImage {
+    const width = image.width
+    const height = image.height
+
+    let bin = this.findBestBin(width, height)
+
+    if (!bin) {
+      bin = {
+        x: 0,
+        y: this.currentHeight,
+        width: this.maxWidth,
+        height,
+      }
+      this.currentHeight += height
+      this.bins.push(bin)
+    }
+
+    const packedImage: PackedImage = {
+      ...image,
+      x: bin.x,
+      y: bin.y,
+    }
+
+    this.splitBin(bin, width, height)
+
+    return packedImage
+  }
+
+  private splitBin(bin: Rect, width: number, height: number): void {
+    const binIndex = this.bins.indexOf(bin)
+    if (binIndex !== -1) {
+      this.bins.splice(binIndex, 1)
+    }
+
+    if (bin.width > width) {
+      this.bins.push({
+        x: bin.x + width,
+        y: bin.y,
+        width: bin.width - width,
+        height: bin.height,
+      })
+    }
+
+    if (bin.height > height) {
+      this.bins.push({
+        x: bin.x,
+        y: bin.y + height,
+        width,
+        height: bin.height - height,
+      })
+    }
+
+    this.currentHeight = Math.max(this.currentHeight, bin.y + height)
+  }
+
+  getHeight(): number {
+    return this.currentHeight
+  }
+}
+
+// Status update function
+async function updateStatus(
+  progress: number,
+  message: string,
+  status: string = 'in_progress',
+): Promise<void> {
   const storage = useStorage('atlas')
   await storage.setItem('status', {
     status,
[…]
