### 2025-03-31 d65961e
build(deps): bump keras from 3.9.0 to 3.9.1
Bumps [keras](https://github.com/keras-team/keras) from 3.9.0 to 3.9.1.
- [Release notes](https://github.com/keras-team/keras/releases)
- [Commits](https://github.com/keras-team/keras/compare/v3.9.0...v3.9.1)

---
updated-dependencies:
- dependency-name: keras
  dependency-type: direct:production
  update-type: version-update:semver-patch
...

Signed-off-by: dependabot[bot] <support@github.com>
