### 2025-04-07 5132313
build(deps): bump keras from 3.9.0 to 3.9.2
Bumps [keras](https://github.com/keras-team/keras) from 3.9.0 to 3.9.2.
- [Release notes](https://github.com/keras-team/keras/releases)
- [Commits](https://github.com/keras-team/keras/compare/v3.9.0...v3.9.2)

---
updated-dependencies:
- dependency-name: keras
  dependency-version: 3.9.2
  dependency-type: direct:production
  update-type: version-update:semver-patch
...

Signed-off-by: dependabot[bot] <support@github.com>
