### 2025-04-07 cd37956
fix: set alpha to 0 for atlas background
