### 2025-04-07 0c2be29
chore(release): 1.0.0-dev.5 [skip ci]
## [1.0.0-dev.5](https://github.com/alexv710/daedalusData/compare/v1.0.0-dev.4...v1.0.0-dev.5) (2025-04-07)

### Features

* add size constraint to atlas generation and improve compatibility ([dabb4a1](https://github.com/alexv710/daedalusData/commit/dabb4a157108510f459aac8c7ed5b31cfd1b8305))
* bump version ([f9d1e89](https://github.com/alexv710/daedalusData/commit/f9d1e894dd157c96f67c82db5c453ad00beb6a3e))
* decouple atlas generation from metadata existance ([15f68ff](https://github.com/alexv710/daedalusData/commit/15f68ff5e94e1534eb514c989fb70a57e449ac6f)
[…]
