+++ frontend/CHANGELOG.md
@@ -3,6 +3,20 @@
 ### Features
 
 * add size constraint to atlas generation and improve compatibility ([dabb4a1](https://github.com/alexv710/daedalusData/commit/dabb4a157108510f459aac8c7ed5b31cfd1b8305))
+* bump version ([f9d1e89](https://github.com/alexv710/daedalusData/commit/f9d1e894dd157c96f67c82db5c453ad00beb6a3e))
+* decouple atlas generation from metadata existance ([15f68ff](https://github.com/alexv710/daedalusData/commit/15f68ff5e94e1534eb514c989fb70a57e449ac6f))
+* fetch all files via a server route ([e1b74a5](https://github.com/alexv710/daedalusData/commit/e1b74a52129abc1ce66feb8d687986b2e08115d6))
+
+### Bug Fixes
+
+* ignore the semantic versioning commits in the commitlint and lint js ([6d07687](https://github.com/alexv710/daedalusData/commit/6d07687633c2e0bef5ce0db852afffc8cf9aaeb2))
+* set alpha to 0 for atlas background ([cd37956](https://github.com/alexv710/daedalusData/commit/cd3795601c663ad8a4f63406fe323e5424740e91))
+
+## [1.0.0-dev.5](https://github.com/alexv710/daedalusData/compare/v1.0.0-dev.4...v1.0.0-dev.5) (2025-04-07)
+
+### Features
+
+* add size constraint to atlas generation and improve compatibility ([dabb4a1](https://github.com/alexv710/daedalusData/commit/dabb4a157108510f459aac8c7ed5b31cfd1b8305))
 * decouple atlas generation from metadata existance ([15f68ff](https://github.com/alexv710/daedalusData/commit/15f68ff5e94e1534eb514c989fb70a57e449ac6f))
 * fetch all files via a server route ([e1b74a5](https://github.com/alexv710/daedalusData/commit/e1b74a52129abc1ce66feb8d687986b2e08115d6))
 
