### 2025-04-17 b60f58f
refactor: optimize the atlas generation using webp and improved scaling
