+++ frontend/server/api/atlasGenerator.ts
@@ -1,545 +1,527 @@
-import type { Sharp } from 'sharp'
-import fs from 'node:fs'
+import nfs from 'node:fs'
+import fs from 'node:fs/promises'
 import path from 'node:path'
-import process from 'node:process'
 import { useStorage } from '#imports'
-import { defineEventHandler } from 'h3'
+import { MaxRectsPacker } from 'maxrects-packer'
 import sharp from 'sharp'
+import pLimit from 'p-limit'
 
-// Constants
-const MAX_ATLAS_WIDTH = 8192
-const MAX_ATLAS_HEIGHT = 8192
-const MAX_PIXEL_COUNT = 67108864
-const METADATA_BATCH_SIZE = 2
-const RESIZE_BATCH_SIZE = 2
-const ALLOWED_EXTENSIONS = ['.png', '.jpg', '.jpeg']
+// --- Configuration ---
+const SHARP_CONCURRENCY = 8
+const ALLOWED_EXTENSIONS = ['.png', '.jpg', '.jpeg', '.webp', '.gif', '.avif', '.tiff']
+// maximum size for webp
+const MAX_ATLAS_DIMENSION = 16383
+const OUTPUT_FORMAT = 'webp'
+const ATLAS_IMAGE_NAME = `atlas.${OUTPUT_FORMAT}`
+const ATLAS_JSON_NAME = 'atlas.json'
+// --- End Configuration ---
+
+// Simple interface for image data needed during processing
+interface ImageProcessingInfo {
+    path: string
+    id: string
+    originalWidth: number
+    originalHeight: number
+    width: number
+    height: number
+    needsResize: boolean
+    // Properties added by the packer
+    x?: number
+    y?: number
+}
+
+interface InitialMetadata {
+    filePath: string
+    filename: string
+    originalWidth: number
+    originalHeight: number
+    largestDim: number
+}
+
+// Simple interface for the final JSON data structure per image
+interface AtlasCoordinates {
+    x: number
+    y: number
+    width: number
+    height: number
+}
 
 // Interfaces
 interface AtlasStatus {
-  status: string
-  progress: number
-  message: string
-  lastUpdated: string
+    status: string
+    progress: number
+    message: string
+    lastUpdated: string
 }
 
-interface ImageMeta {
-  filename: string
-  width: number
-  height: number
-  originalWidth: number
-  originalHeight: number
-}
-
-interface PackedImage extends ImageMeta {
-  x: number
-  y: number
-}
-
-interface Rect {
-  x: number
-  y: number
-  width: number
-  height: number
-}
-
-// Bin-packing algorithm
-class Packer {
-  private bins: Rect[] = []
-  private readonly maxWidth: number
-  private currentHeight: number = 0
-
-  constructor(maxWidth: number) {
-    this.maxWidth = maxWidth
-    this.bins.push({ x: 0, y: 0, width: maxWidth, height: 0 })
-  }
-
-  private findBestBin(width: number, height: number): Rect | null {
-    let bestBin: Rect | null = null
-    let bestScore = Infinity
-
-    for (const bin of this.bins) {
-      if (bin.width < width || bin.height < height)
-        continue
-
-      const score = bin.width - width
-      if (score < bestScore) {
-        bestScore = score
-        bestBin = bin
-      }
-    }
-
-    return bestBin
-  }
-
-  pack(image: ImageMeta): PackedImage {
-    const width = image.width
-    const height = image.height
-
-    let bin = this.findBestBin(width, height)
-
-    if (!bin) {
-      bin = {
-        x: 0,
-        y: this.currentHeight,
-        width: this.maxWidth,
-        height,
-      }
-      this.currentHeight += height
[…]
