### 2025-04-19 9626b48
refactor: improve the filestream retreival from the api
