+++ frontend/server/api/file/[...params].get.ts
@@ -1,124 +1,102 @@
-import fs, { createReadStream } from 'node:fs'
+import fs from 'node:fs'
 import path from 'node:path'
 import { createError, defineEventHandler, sendStream } from 'h3'
 
-export default defineEventHandler(async (event) => {
-  // Retrieve the catch-all parameter.
-  // When using a catch-all route (e.g. [ ...params ].get.ts),
-  // the parameter might be passed as a string (with slashes) or as an array.
-  const rawParams = event.context.params?.params
-  let params: string[] = []
-  if (typeof rawParams === 'string') {
-    params = rawParams.split('/')
-  }
-  else if (Array.isArray(rawParams)) {
-    params = rawParams
-  }
-  console.info('Requested params:', params)
-  console.info('Request method:', event.node.req.method)
 
-  let relativePath: string
-  if (params.length === 1) {
-    // For a single parameter, if it's "atlas.png" or "atlas.json", serve from the root.
-    if (params[0] === 'atlas.png' || params[0] === 'atlas.json') {
-      relativePath = params[0]
-    }
-    else {
-      // Otherwise, assume it's an image in the "images" folder.
-      relativePath = path.join('images', params[0])
-    }
+const CONTENT_TYPE_MAP: Record<string, string> = {
+  '.png': 'image/png',
+  '.webp': 'image/webp',
+  '.jpg': 'image/jpeg',
+  '.jpeg': 'image/jpeg',
+  '.gif': 'image/gif',
+  '.json': 'application/json',
+};
+const DEFAULT_CONTENT_TYPE = 'application/octet-stream';
+
+async function findRelevantPaths(): Promise<string> {
+  const basePaths = [
+    path.join(process.cwd(), '/app/data'),
+    path.join(process.cwd(), '/data'),
+    path.join(process.cwd(), '../data'),
+  ]
+  
+  const dataDir = basePaths.find(p => fs.existsSync(p))
+  if (!dataDir) {
+    throw new Error('Could not find an accessible data directory')
   }
-  else if (params.length === 2) {
-    const dir = params[0].toLowerCase()
-    const file = params[1]
-    // Allowed directories.
-    const allowedDirs = ['images', 'metadata', 'labels', 'projections', 'features']
-    if (!allowedDirs.includes(dir)) {
-      throw createError({
-        statusCode: 400,
-        statusMessage: 'Bad Request',
-        message: 'Invalid directory requested',
-      })
-    }
-    // For directories other than "images", only allow .json files.
-    if (dir !== 'images' && !file.toLowerCase().endsWith('.json')) {
-      throw createError({
-        statusCode: 400,
-        statusMessage: 'Bad Request',
-        message: `Only .json files are allowed in the ${dir} directory`,
-      })
-    }
-    relativePath = path.join(dir, file)
-  }
-  else {
+  return dataDir
+}
+const DATA_ROOT = await findRelevantPaths()
+console.info(`Serving static files from: ${DATA_ROOT}`);
+
+export default defineEventHandler(async (event) => {
+  const rawParams = event.context.params?.params;
+  const requestedRelativePath = Array.isArray(rawParams) ? rawParams.join('/') : rawParams || '';
+
+  if (!requestedRelativePath) {
     throw createError({
       statusCode: 400,
       statusMessage: 'Bad Request',
-      message: 'Invalid parameters',
-    })
+      message: 'No file path specified in the request.',
+    });
   }
 
-  // Define possible base directories. This attempts /app/data first and falls back to /data.
-  const basePaths = [
-    path.join(process.cwd(), 'app/data'),
-    path.join(process.cwd(), 'data'),
-    path.join(process.cwd(), '../data'),
-  ]
[…]
