### 2025-04-19 952877e
refactor: improve the check for the right data directory
