### 2025-04-19 db4d013
fix: remove the unocss font, as fetching it always failed
