### 2025-04-19 7e6470b
test: add test script to generate test image dataset with numbers
