### 2025-04-19 1f7972c
style: resolve linting and formatting issues
