+++ frontend/server/api/atlasGenerator.ts
@@ -1,10 +1,11 @@
+import type { Buffer } from 'node:buffer'
 import nfs from 'node:fs'
 import fs from 'node:fs/promises'
 import path from 'node:path'
 import { useStorage } from '#imports'
 import { MaxRectsPacker } from 'maxrects-packer'
-import sharp from 'sharp'
 import pLimit from 'p-limit'
+import sharp from 'sharp'
 
 // --- Configuration ---
 const SHARP_CONCURRENCY = 8
@@ -18,95 +19,95 @@ const ATLAS_JSON_NAME = 'atlas.json'
 
 // Simple interface for image data needed during processing
 interface ImageProcessingInfo {
-    path: string
-    id: string
-    originalWidth: number
-    originalHeight: number
-    width: number
-    height: number
-    needsResize: boolean
-    // Properties added by the packer
-    x?: number
-    y?: number
+  path: string
+  id: string
+  originalWidth: number
+  originalHeight: number
+  width: number
+  height: number
+  needsResize: boolean
+  // Properties added by the packer
+  x?: number
+  y?: number
 }
 
 interface InitialMetadata {
-    filePath: string
-    filename: string
-    originalWidth: number
-    originalHeight: number
-    largestDim: number
+  filePath: string
+  filename: string
+  originalWidth: number
+  originalHeight: number
+  largestDim: number
 }
 
 // Simple interface for the final JSON data structure per image
 interface AtlasCoordinates {
-    x: number
-    y: number
-    width: number
-    height: number
+  x: number
+  y: number
+  width: number
+  height: number
 }
 
 // Interfaces
 interface AtlasStatus {
-    status: string
-    progress: number
-    message: string
-    lastUpdated: string
+  status: string
+  progress: number
+  message: string
+  lastUpdated: string
 }
 
 // Dummy updateStatus function (replace with your actual implementation)
 async function updateStatus(
-    progress: number,
-    message: string,
-    status: string = 'in_progress',
+  progress: number,
+  message: string,
+  status: string = 'in_progress',
 ): Promise<void> {
-    const storage = useStorage('atlas')
-    await storage.setItem('status', {
-        status,
-        progress,
-        message,
-        lastUpdated: new Date().toISOString(),
-    } as AtlasStatus)
+  const storage = useStorage('atlas')
+  await storage.setItem('status', {
+    status,
+    progress,
+    message,
+    lastUpdated: new Date().toISOString(),
+  } as AtlasStatus)
 }
 
 /**
  * Finds base data dir, depending o
  */
 async function findRelevantPaths(): Promise<string> {
-    const basePaths = [
-        path.join(process.cwd(), '/app/data'),
-        path.join(process.cwd(), '/data'),
-        path.join(process.cwd(), '../data'),
-    ]
+  const basePaths = [
+    path.join(process.cwd(), '/app/data'),
+    path.join(process.cwd(), '/data'),
+    path.join(process.cwd(), '../data'),
+  ]
 
-    const dataDir = basePaths.find(p => nfs.existsSync(p))
-    if (!dataDir) {
-        throw new Error('Could not find an accessible data directory')
-    }
-    return dataDir
+  const dataDir = basePaths.find(p => nfs.existsSync(p))
+  if (!dataDir) {
+    throw new Error('Could not find an accessible data directory')
+  }
+  return dataDir
 }
 
 /**
  * Scans a directory for image files with allowed extensions.
  */
 async function getImageFilePaths(imagesDir: string): Promise<string[]> {
-    let files: string[]
-    try {
-        files = await fs.readdir(imagesDir)
-    }
-    catch (err) {
-        throw new Error(`Images directory not found or could not be read at ${imagesDir}: ${err.message}`)
-    }
+  let files: string[]
+  try {
[…]
