### 2025-04-19 61a7c9f
build: update github actions for the py safety check
