### 2025-04-28 6e30fe2
build(deps): bump torchvision from 0.21.0 to 0.22.0
Bumps [torchvision](https://github.com/pytorch/vision) from 0.21.0 to 0.22.0.
- [Release notes](https://github.com/pytorch/vision/releases)
- [Commits](https://github.com/pytorch/vision/compare/v0.21.0...v0.22.0)

---
updated-dependencies:
- dependency-name: torchvision
  dependency-version: 0.22.0
  dependency-type: direct:production
  update-type: version-update:semver-minor
...

Signed-off-by: dependabot[bot] <support@github.com>
