### 2025-04-28 844c3a4
build(deps): bump pyarrow from 19.0.1 to 20.0.0
Bumps [pyarrow](https://github.com/apache/arrow) from 19.0.1 to 20.0.0.
- [Release notes](https://github.com/apache/arrow/releases)
- [Commits](https://github.com/apache/arrow/compare/apache-arrow-19.0.1...apache-arrow-20.0.0)

---
updated-dependencies:
- dependency-name: pyarrow
  dependency-version: 20.0.0
  dependency-type: direct:production
  update-type: version-update:semver-major
...

Signed-off-by: dependabot[bot] <support@github.com>
