### 2025-05-12 0173470
build(deps): bump node from 20-slim to 24-slim
Bumps node from 20-slim to 24-slim.

---
updated-dependencies:
- dependency-name: node
  dependency-version: 24-slim
  dependency-type: direct:production
...

Signed-off-by: dependabot[bot] <support@github.com>
