+++ requirements.txt
@@ -1,5 +1,5 @@
 
-keras==3.9.0
+keras==3.10.0
 numpy==2.0.2
 pandas==2.2.3
 pillow==11.1.0
