### 2025-06-01 #8325 @editorialbot
[REVIEW]: DaedalusData: A Dockerized Platform for Exploration, Knowledge Externalization and Labeling of Image Collections
**Submitting author:** @alexv710 (<a href="http://orcid.org/0009-0009-2763-3186">Alexander Wyss</a>)
**Repository:** https://github.com/alexv710/daedalusData
**Branch with paper.md** (empty if default branch): fb-paper
**Version:** v1.0.0
