### 2025-06-16 59ba6a4
build(deps): bump tensorflow from 2.18.0 to 2.19.0
Bumps [tensorflow](https://github.com/tensorflow/tensorflow) from 2.18.0 to 2.19.0.
- [Release notes](https://github.com/tensorflow/tensorflow/releases)
- [Changelog](https://github.com/tensorflow/tensorflow/blob/master/RELEASE.md)
- [Commits](https://github.com/tensorflow/tensorflow/compare/v2.18.0...v2.19.0)

---
updated-dependencies:
- dependency-name: tensorflow
  dependency-version: 2.19.0
  dependency-type: direct:production
  update-type: version-update:semver-minor
...

Signed-off-by: dependabot[bot] <support@github.com>
