+++ requirements.txt
@@ -5,7 +5,7 @@ pandas==2.2.3
 pillow==11.1.0
 scikit-learn==1.6.1
 umap-learn==0.5.7
-tensorflow==2.18.0
+tensorflow==2.19.0
 torchvision==0.21.0
 pyarrow==19.0.1
 pymysql==1.1.1
