### 2025-07-07 81234b7
build(deps): bump umap-learn from 0.5.7 to 0.5.9.post2
Bumps [umap-learn](https://github.com/lmcinnes/umap) from 0.5.7 to 0.5.9.post2.
- [Release notes](https://github.com/lmcinnes/umap/releases)
- [Changelog](https://github.com/lmcinnes/umap/blob/master/doc/release_notes.rst)
- [Commits](https://github.com/lmcinnes/umap/compare/release-0.5.7...release-0.5.9.post2)

---
updated-dependencies:
- dependency-name: umap-learn
  dependency-version: 0.5.9.post2
  dependency-type: direct:production
  update-type: version-update:semver-patch
...

Signed-off-by: dependabot[bot] <support@github.com>
