+++ requirements.txt
@@ -4,7 +4,7 @@ numpy==2.0.2
 pandas==2.2.3
 pillow==11.1.0
 scikit-learn==1.6.1
-umap-learn==0.5.7
+umap-learn==0.5.9.post2
 tensorflow==2.18.0
 torchvision==0.21.0
 pyarrow==19.0.1
