### 2025-07-13 #8325 @sara-02
Hello folks,
I am working on the review, will open the issues on the project for
respective shortcomings.

On Thu, 10 Jul 2025 at 04:45, George K. Thiruvathukal <
***@***.***> wrote:

> *gkthiruvathukal* left a comment (openjournals/joss-reviews#8325)
> <https://github.com/openjournals/joss-reviews/issues/8325#issuecomment-3055137104>
>
> @tushardave26 <https://github.com/tushardave26> and @sara-02
> <https://github.com/sara-02>: Can you please provide an update on your
> review?
>
