### 2025-08-11 49a52cf
build(deps): bump torchvision from 0.21.0 to 0.23.0
Bumps [torchvision](https://github.com/pytorch/vision) from 0.21.0 to 0.23.0.
- [Release notes](https://github.com/pytorch/vision/releases)
- [Commits](https://github.com/pytorch/vision/compare/v0.21.0...0.23.0)

---
updated-dependencies:
- dependency-name: torchvision
  dependency-version: 0.23.0
  dependency-type: direct:production
  update-type: version-update:semver-minor
...

Signed-off-by: dependabot[bot] <support@github.com>
