+++ requirements.txt
@@ -6,7 +6,7 @@ pillow==11.1.0
 scikit-learn==1.6.1
 umap-learn==0.5.7
 tensorflow==2.18.0
-torchvision==0.21.0
+torchvision==0.23.0
 pyarrow==19.0.1
 pymysql==1.1.1
 jinja2==3.1.6
