### 2025-08-25 bf28e50
build(deps): bump pymysql from 1.1.1 to 1.1.2
Bumps [pymysql](https://github.com/PyMySQL/PyMySQL) from 1.1.1 to 1.1.2.
- [Release notes](https://github.com/PyMySQL/PyMySQL/releases)
- [Changelog](https://github.com/PyMySQL/PyMySQL/blob/main/CHANGELOG.md)
- [Commits](https://github.com/PyMySQL/PyMySQL/compare/v1.1.1...v1.1.2)

---
updated-dependencies:
- dependency-name: pymysql
  dependency-version: 1.1.2
  dependency-type: direct:production
  update-type: version-update:semver-patch
...

Signed-off-by: dependabot[bot] <support@github.com>
