+++ requirements.txt
@@ -8,5 +8,5 @@ umap-learn==0.5.7
 tensorflow==2.18.0
 torchvision==0.21.0
 pyarrow==19.0.1
-pymysql==1.1.1
+pymysql==1.1.2
 jinja2==3.1.6
