### 2025-09-01 alexv710/daedalusData#142 @tushardave26
Post-Docker Instructions
Hello @alexv710,

This issue is a part of the JOSS review for your submission (https://github.com/openjournals/joss-reviews/issues/8325).

Once the container is up and running with `docker compose up -d`, it’s unclear what the next step is. A sentence or two in the `README` describing what users should expect to see and where to navigate in the browser would improve the onboarding experience.

Thanks,
Tushar
