### 2025-09-01 alexv710/daedalusData#143 @tushardave26
FileNotFoundError in Dimensionality_Reduction.ipynb Example Notebook
Hello @alexv710,

This issue is a part of the JOSS review for your submission (https://github.com/openjournals/joss-reviews/issues/8325).

While going through the `Dimensionality_Reduction.ipynb` notebook as part of my [JOSS review](https://github.com/openjournals/joss-reviews/issues/8325?utm_source=chatgpt.com), I encountered the following error when running the example code:

```
FileNotFoundError: [Errno 2] No such file or directory: '/app/data/features/image_features.npz'
```

Please address the issue and let me know once it's resolved so I can re-test the code.

Thanks.
