### 2025-09-01 alexv710/daedalusData#144 @tushardave26
Review comments from JOSS Reviewer (@tushardave26)
Hi @alexv710 👋 ,

Thanks for submitting DaedalusData to JOSS. I've begun reviewing the repository and manuscript, and overall, the structure and goals of the project are clear and aligned with JOSS's scope.

As part of my review of the JOSS submission (https://github.com/openjournals/joss-reviews/issues/8325), I’ll be opening GitHub issues to provide structured and actionable feedback. Please review and address each issue individually. Once you’ve responded to an item, I’ll revisit it and update the corresponding checkboxes below.

- [ ] #142
- [ ] #143

I'll continue to test the functionality and review the codebase more thoroughly. I’ll raise any follow-up suggestions as GitHub issues directly on the repository.

Thanks again for sharing this work.
