### 2025-10-06 6eecec4
build(deps): bump pandas from 2.2.3 to 2.3.3
Bumps [pandas](https://github.com/pandas-dev/pandas) from 2.2.3 to 2.3.3.
- [Release notes](https://github.com/pandas-dev/pandas/releases)
- [Commits](https://github.com/pandas-dev/pandas/compare/v2.2.3...v2.3.3)

---
updated-dependencies:
- dependency-name: pandas
  dependency-version: 2.3.3
  dependency-type: direct:production
  update-type: version-update:semver-minor
...

Signed-off-by: dependabot[bot] <support@github.com>
