### 2025-10-13 aeb8a3a
build(deps): bump keras from 3.9.0 to 3.11.3
Bumps [keras](https://github.com/keras-team/keras) from 3.9.0 to 3.11.3.
- [Release notes](https://github.com/keras-team/keras/releases)
- [Commits](https://github.com/keras-team/keras/compare/v3.9.0...v3.11.3)

---
updated-dependencies:
- dependency-name: keras
  dependency-version: 3.11.3
  dependency-type: direct:production
  update-type: version-update:semver-minor
...

Signed-off-by: dependabot[bot] <support@github.com>
