+++ requirements.txt
@@ -1,6 +1,6 @@
 
 keras==3.9.0
-numpy==2.0.2
+numpy==2.3.4
 pandas==2.2.3
 pillow==11.1.0
 scikit-learn==1.6.1
