+++ Dockerfile
@@ -1,4 +1,4 @@
-FROM node:20-slim as dev
+FROM node:25-slim as dev
 WORKDIR /app
 
 # Install pnpm with minimal cache
