+++ requirements.txt
@@ -7,6 +7,6 @@ scikit-learn==1.6.1
 umap-learn==0.5.7
 tensorflow==2.18.0
 torchvision==0.21.0
-pyarrow==19.0.1
+pyarrow==22.0.0
 pymysql==1.1.1
 jinja2==3.1.6
