### 2025-11-03 79273f6
build(deps): bump keras from 3.9.0 to 3.12.0
Bumps [keras](https://github.com/keras-team/keras) from 3.9.0 to 3.12.0.
- [Release notes](https://github.com/keras-team/keras/releases)
- [Commits](https://github.com/keras-team/keras/compare/v3.9.0...v3.12.0)

---
updated-dependencies:
- dependency-name: keras
  dependency-version: 3.12.0
  dependency-type: direct:production
  update-type: version-update:semver-minor
...

Signed-off-by: dependabot[bot] <support@github.com>
