### 2026-08-21 #8325 @matt-graham
:wave: @itepifanio, @elboyran, @drcandacemakedamoore, would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html. In this case the review has already started but as we've not had recent contact from one of the reviewers we are looking for an additional reviewer. Please let me know even if you're not interested, so that I can look for other reviewers.
