### 2026-08-21 #8325 @matt-graham
Thanks @drcandacemakedamoore! Once you are ready to start your review you can create your checklist by entering

```@editorialbot generate my checklist```
