### 2026-08-21 #8325 @drcandacemakedamoore
@matt-graham  I have an immediate comment that goes outside the checklist. The title is very confusing in light of related work with the same title (https://ieeexplore.ieee.org/abstract/document/10689477) which was already published that I believe to be the same team. Do we need to check that there is sufficient innovation for this to be a new software package, and therefore needing another publication? ALso if it was building on the previous work, what about the previous authors.... also, to the authors, the about section on your Github is blank.
