### 2026-08-24 #8325 @drcandacemakedamoore
Thank you @alexv710 . I have been working on a Windows machine lately, and I can't get the docker to work.  I tinkered with the code, and also calling docker with slightly different commands, but to no avail. Maybe this was only meant for Linux? Can you add , if I just have not found it yet a matrix of systems this thing has been tested to work on? I tend to put this right on the README of many projects (as an example https://github.com/drcandacemakedamoore/cleanX support platforms section) but it should be somewhere not too hidden. Apologies if I missed this.
