### 2026-08-25 #8325 @matt-graham
> @matt-graham I have an immediate comment that goes outside the checklist. The title is very confusing in light of related work with the same title (https://ieeexplore.ieee.org/abstract/document/10689477) which was already published that I believe to be the same team. They acklowledge it so this is not malicious. But do we need to check that there is sufficient innovation for this to be a new software package, and therefore needing another publication?

Thanks @drcandacemakedamoore for raising this point with regards to the related publication. In general JOSS [allows co-publishing papers which describe a software package alongside a separate contribution describing for example a science application or methods development](https://joss.readthedocs.io/en/latest/submitting.html#co-publication-of-science-methods-and-software).
