### 2026-08-30 #8325 @alexv710
Hi @drcandacemakedamoore

Thanks for catching this, and sorry for the trouble. Your report led me to two real bugs. First, a recent pnpm release broke the image build entirely: the Dockerfile installed an unpinned pnpm, and pnpm 10 rejects a config command used during the build. This failed on any OS with a fresh clone, so it is almost certainly what you ran into. Second, a fresh clone was missing the data/ directory and the compose file still mounted a directory that no longer exists, Linux silently creates missing bind mount sources but Docker Desktop on Windows fails on them. Both are fixed in the latest release, and I verified a fresh clone now starts cleanly.

I also added a "Supported Platforms" section near the top of the README, along the lines of the cleanX one you linked. On Windows, Docker Desktop needs to be in Linux container mode with the WSL 2 backend, ideally with the repo cloned inside the WSL filesystem, that is covered in the new platform notes.

If it still fails after pulling the latest main, the error message and your Docker version would help a lot.

Thanks and Best, Alex
